feat: CustomKeywordView 재사용 가능하게 클래스로 분리
