+++ TILApp/Views/MyProfileViewController/BlogEditViewController.swift
@@ -17,7 +17,7 @@ final class BlogEditViewController: UIViewController {
     }
 
     private lazy var contentView = UIView()
-    private lazy var rootFlexContainer = UIView()
+    private lazy var customKeywordView = CustomKeywordView()
 
     private lazy var mainBlogButton = CustomLargeBorderButton().then {
         $0.variant = blog.main ? .primary : .normal
@@ -126,7 +126,7 @@ final class BlogEditViewController: UIViewController {
 
         view.addSubview(contentScrollView)
         contentScrollView.addSubview(contentView)
-        contentView.addSubview(rootFlexContainer)
+        contentView.addSubview(customKeywordView)
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -171,26 +171,11 @@ final class BlogEditViewController: UIViewController {
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
 
-        rootFlexContainer.removeAllSubviews()
-
-        rootFlexContainer.flex.define {
-            for (index, keyword) in keywordInputViewModel.keywords.enumerated() {
-                let customTagView = CustomTagView()
-                customTagView.labelText = keyword.keyword
-                customTagView.tags = keyword.tags
-                $0.addItem(customTagView).marginTop(10)
-                customTagView.pin.size(customTagView.componentSize)
-
-                let tapGestureRecognizer
-                    = ContextTapGestureRecognizer(target: self, action: #selector(customTagViewTapped(_:)))
-                tapGestureRecognizer.context["index"] = index
-                customTagView.addGestureRecognizer(tapGestureRecognizer)
-                customTagView.isUserInteractionEnabled = true
-
-                customTagView
-                    .addTargetForButton(target: self, action: #selector(deleteTagButtonTapped), for: .touchUpInside)
-            }
-        }
+        customKeywordView.setKeywords(
+            keywordInputViewModel.keywords,
+            target: self, tapSelector: #selector(customTagViewTapped(_:)),
+            deleteSelector: #selector(deleteTagButtonTapped(_:))
+        )
 
         contentScrollView.pin.top(view.pin.safeArea).horizontally().bottom()
         contentView.pin.top(to: contentScrollView.edge.top).horizontally()
@@ -204,10 +189,10 @@ final class BlogEditViewController: UIViewController {
 
         tagHeader.pin.top(to: deleteBlogButton.edge.bottom).marginTop(20)
 
-        rootFlexContainer.pin.horizontally(20).top(to: tagHeader.edge.bottom).bottom().marginTop(-10)
-        rootFlexContainer.flex.layout(mode: .adjustHeight)
+        customKeywordView.pin.horizontally(20).top(to: tagHeader.edge.bottom).bottom().marginTop(-10)
+        customKeywordView.flex.layout(mode: .adjustHeight)
 
-        contentView.pin.top(to: contentScrollView.edge.top).horizontally().bottom(to: rootFlexContainer.edge.bottom)
+        contentView.pin.top(to: contentScrollView.edge.top).horizontally().bottom(to: customKeywordView.edge.bottom)
 
         contentScrollView.contentSize = CGSize(width: contentView.frame.width, height: contentView.frame.height)
     }
@@ -229,7 +214,7 @@ final class BlogEditViewController: UIViewController {
 
     @objc private func deleteTagButtonTapped(_ sender: UIButton) {
         if let customTagView = sender.superview as? CustomTagView,
-           let index = rootFlexContainer.subviews.firstIndex(of: customTagView)
+           let index = customKeywordView.subviews.firstIndex(of: customTagView)
         {
             let keyword = keywordInputViewModel.keywords[index]
             let alertController = UIAlertController(
