+++ TILApp/Views/MyProfileViewController/BlogRegisterViewController.swift
@@ -46,7 +46,7 @@ final class BlogRegisterViewController: UIViewController, UIGestureRecognizerDel
         navigationController?.pushViewController(EditTagViewController(), animated: true)
     }
 
-    private lazy var rootFlexContainer = UIView()
+    private lazy var customKeywordView = CustomKeywordView()
 
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -60,7 +60,7 @@ final class BlogRegisterViewController: UIViewController, UIGestureRecognizerDel
 
         view.addSubview(contentScrollView)
         contentScrollView.addSubview(contentView)
-        contentView.addSubview(rootFlexContainer)
+        contentView.addSubview(customKeywordView)
     }
 
     // TODO: 프린트문 삭제
@@ -92,26 +92,11 @@ final class BlogRegisterViewController: UIViewController, UIGestureRecognizerDel
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
 
-        rootFlexContainer.removeAllSubviews()
-
-        rootFlexContainer.flex.define {
-            for (index, keyword) in keywordInputViewModel.keywords.enumerated() {
-                let customTagView = CustomTagView()
-                customTagView.labelText = keyword.keyword
-                customTagView.tags = keyword.tags
-                $0.addItem(customTagView).marginTop(10)
-                customTagView.pin.size(customTagView.componentSize)
-
-                let tapGestureRecognizer
-                    = ContextTapGestureRecognizer(target: self, action: #selector(customTagViewTapped(_:)))
-                tapGestureRecognizer.context["index"] = index
-                customTagView.addGestureRecognizer(tapGestureRecognizer)
-                customTagView.isUserInteractionEnabled = true
-
-                customTagView
-                    .addTargetForButton(target: self, action: #selector(deleteTagButtonTapped), for: .touchUpInside)
-            }
-        }
+        customKeywordView.setKeywords(
+            keywordInputViewModel.keywords,
+            target: self, tapSelector: #selector(customTagViewTapped(_:)),
+            deleteSelector: #selector(deleteTagButtonTapped(_:))
+        )
 
         contentScrollView.pin.top(view.pin.safeArea).horizontally().bottom()
         contentView.pin.top(to: contentScrollView.edge.top).horizontally()
@@ -121,10 +106,10 @@ final class BlogRegisterViewController: UIViewController, UIGestureRecognizerDel
         blogRSSTextField.pin.horizontally().top(to: blogURLTextField.edge.bottom).marginTop(5)
         tagHeader.pin.top(to: blogRSSTextField.edge.bottom).marginTop(5)
 
-        rootFlexContainer.pin.horizontally(20).top(to: tagHeader.edge.bottom).bottom().marginTop(-10)
-        rootFlexContainer.flex.layout(mode: .adjustHeight)
+        customKeywordView.pin.horizontally(20).top(to: tagHeader.edge.bottom).bottom().marginTop(-10)
+        customKeywordView.flex.layout(mode: .adjustHeight)
 
-        contentView.pin.top(to: contentScrollView.edge.top).horizontally().bottom(to: rootFlexContainer.edge.bottom)
+        contentView.pin.top(to: contentScrollView.edge.top).horizontally().bottom(to: customKeywordView.edge.bottom)
 
         contentScrollView.contentSize = CGSize(width: contentView.frame.width, height: contentView.frame.height)
     }
@@ -142,7 +127,7 @@ final class BlogRegisterViewController: UIViewController, UIGestureRecognizerDel
 
     @objc private func deleteTagButtonTapped(_ sender: UIButton) {
         if let customTagView = sender.superview as? CustomTagView,
-           let index = rootFlexContainer.subviews.firstIndex(of: customTagView)
+           let index = customKeywordView.subviews.firstIndex(of: customTagView)
         {
             let keyword = keywordInputViewModel.keywords[index]
             let alertController = UIAlertController(
